requirements.txt
asgiref==3.7.2
coverage==7.3.2
Django==4.2.6
django-filter==23.3
gunicorn==21.2.0
packaging==23.2
Pillow==10.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0

reportlab~=4.0.7
cffi~=1.16.0
pip~=23.3.1
wheel~=0.41.2
cryptography~=41.0.5
setuptools~=68.2.2
PyYAML~=6.0.1
requests~=2.31.0